Refactor Size() on entities to return relative units
Previously, Size() returned an absolute pixel value, which made it quite
inflexible, especially when pixel values are not used in any other part of
the application. By using relative floating point numbers instead, the
process is much simpler and more effective.

+++ server/entities/entity.go
@@ -36,7 +36,7 @@ type Entity interface {
 
     ID() string
     Position() (float64, float64)
-    Size() (uint, uint)
+    Size() (float64, float64)
 
     Location() EntityRegion
     Inventory() *Inventory // May return nil
